Add the ability to parse chat messages with tags.
Adding tags to the chat messages required a different way to handle the
incoming messages. To take this into account we added a message struct
that can hold the values of the different fields provided in the tags.
Most of them are not currently used but some of them are nice to add
features. For example, using the color tag we can now color the username
in the output based on what it is on Twitch.